Return stream id on auth, split store.SetActive
This commit make the Auth function return the published streams id.
The reason behind this is simply that this is the only point, where its
possible to gather which stream is actually beeing published, because the
server reveals the publishers authkey at this point.

With this change its also possible to set the streams active flag by its
id. Deactivation on unpublishing can only be done by looking at the
streams Application/Name Pair and thus the SetActive function is split
into SetActive(id) which can is called by the streams id and
SetInactive(app, name) which sets all known streams for that app/name
pair as not active.